# Optimize gpr_refcount and gpr_stats_counter helpers.