This patch introduces three main changes:
1) Make the simple methods inline: The functions are all non-inline, meaning that they all impose an extra call and a barrier for a single instruction.
2) Use gpr_atm_no_barrier_store to initialize ref counts: There is no need to create barriers for initializing an atomic similar to ref.
3) Use GPR_DEBUG_ASSERT instead of GPR_ASSERT to save an expensive branch on optimized builds.

Since include/grpc/support/sync.h is a public C API, we cannot use the inline keyword (i.e., a C89 limiation). Thus, this patch introduces a C++ helper that is used to implement the C API, and also in all internal code.

Perfromance resutls are stellar (1.7% CPU time, 3% throughput) and stable with 30x runs:

Before:
BM_StreamingPingPong<TCP, NoOpMutator, NoOpMutator>/1/1_mean        58714 ns      58687 ns      11776   33.2815kB/s
BM_StreamingPingPong<TCP, NoOpMutator, NoOpMutator>/1/1_median      58585 ns      58562 ns      11776   33.3512kB/s
BM_StreamingPingPong<TCP, NoOpMutator, NoOpMutator>/1/1_stddev        376 ns        369 ns      11776    211.818B/s

After:
BM_StreamingPingPong<TCP, NoOpMutator, NoOpMutator>/1/1_mean        56673 ns      56649 ns      12334   34.4782kBs
BM_StreamingPingPong<TCP, NoOpMutator, NoOpMutator>/1/1_median      56630 ns      56612 ns      12334   34.4999kB/s
BM_StreamingPingPong<TCP, NoOpMutator, NoOpMutator>/1/1_stddev        237 ns        234 ns      12334    144.449B/s